% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{Birkbak.2018.Sig}
\alias{Birkbak.2018.Sig}
\title{Function to compute the signature as published by Birkbak et al. in 2018.}
\usage{
Birkbak.2018.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Birkbak et al. in 2018.
}
\examples{
a = Birkbak.2018.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Birkbak, N J et al. “Overexpression of BLM promotes DNA damage and increased sensitivity to platinum salts in triple-negative breast and serous ovarian cancers.” Annals of oncology : official journal of the European Society for Medical Oncology vol. 29,4 (2018): 903-909. doi:10.1093/annonc/mdy049
}
