% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{BRCAness.Sig}
\alias{BRCAness.Sig}
\title{Function to compute the signature as published by Oshi et al. in 2022.}
\usage{
BRCAness.Sig(expr, Response = NULL, verbose = FALSE, ...)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}

\item{...}{Other arguments passed on to gsva()'s params argument.}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Oshi et al. in 2022.
}
\examples{
a = BRCAness.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Oshi, Masanori et al. “Development of a novel BRCAness score that predicts response to PARP inhibitors.” Biomarker research vol. 10,1 80. 12 Nov. 2022, doi:10.1186/s40364-022-00427-8
}
