% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{calculateSig}
\alias{calculateSig}
\title{Function to compute the signatures.}
\usage{
calculateSig(expr, Response = NULL, Sig_list, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{Sig_list}{The signatures that need to be calculated.Setting the parameter to \code{all} means calculating all signatures.Setting the parameter to \code{Tax} or \code{ICI} means calculating all signatures for the corresponding category.}

\item{verbose}{Gives information about each calculation step.}
}
\description{
This function computes signatures scores from gene expression values.
}
\examples{
file_path = system.file("extdata", "Sig_table.csv", package = "PRTI")
Sig_table = read.csv(file_path,header = TRUE)
Sig_list = Sig_table$Signature[1:5]
res = calculateSig(expr = gene_table, Response = pdata_table$Response, Sig_list = Sig_list)

}
