% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{CD8A.Sig}
\alias{CD8A.Sig}
\title{Function to compute the signature as published by Niu et al. in 2022.}
\usage{
CD8A.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Niu et al. in 2022.
}
\examples{
a = CD8A.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Niu, Decao et al. “The epiphany derived from T-cell-inflamed profiles: Pan-cancer characterization of CD8A as a biomarker spanning clinical relevance, cancer prognosis, immunosuppressive environment, and treatment responses.” Frontiers in genetics vol. 13 974416. 11 Aug. 2022, doi:10.3389/fgene.2022.974416
}
