% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{CES.Sig}
\alias{CES.Sig}
\title{Function to compute the signature as published by Zhang et al. in 2016.}
\usage{
CES.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Zhang et al. in 2016.
}
\examples{
a = CES.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Zhang, Weiguo et al. “Centromere and kinetochore gene misexpression predicts cancer patient survival and response to radiotherapy and chemotherapy.” Nature communications vol. 7 12619. 31 Aug. 2016, doi:10.1038/ncomms12619
}
