% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{CIN70.Sig}
\alias{CIN70.Sig}
\title{Function to compute the signature as published by Carter et al. in 2006.}
\usage{
CIN70.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Carter et al. in 2006.
}
\examples{
a = CIN70.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Carter, Scott L et al. “A signature of chromosomal instability inferred from gene expression profiles predicts clinical outcome in multiple human cancers.” Nature genetics vol. 38,9 (2006): 1043-8. doi:10.1038/ng1861
}
