% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{DDIR.Sig}
\alias{DDIR.Sig}
\title{Function to compute the signature as published by Malla et al. in 2020.}
\usage{
DDIR.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Malla et al. in 2020.
}
\examples{
a = DDIR.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Malla, Sudhir B et al. “In-depth Clinical and Biological Exploration of DNA Damage Immune Response as a Biomarker for Oxaliplatin Use in Colorectal Cancer.” Clinical cancer research : an official journal of the American Association for Cancer Research vol. 27,1 (2021): 288-300. doi:10.1158/1078-0432.CCR-20-3237
}
