% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{Docetaxel.Sig}
\alias{Docetaxel.Sig}
\title{A function for calculating the average value of target genes of docetaxel in DrugBank.}
\usage{
Docetaxel.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
A list of target genes with direct pharmacological action with docetaxel is found from DrugBank, and the average value of their expression is calculated.
}
\examples{
a = Docetaxel.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Knox, Craig et al. “DrugBank 6.0: the DrugBank Knowledgebase for 2024.” Nucleic acids research vol. 52,D1 (2024): D1265-D1275. doi:10.1093/nar/gkad976
}
