% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{GBP2.Sig}
\alias{GBP2.Sig}
\title{Function to compute the signature as published by Wang et al. in 2007.}
\usage{
GBP2.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Wang et al. in 2007.
}
\examples{
a = GBP2.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Wang, Yunfei et al. “GBP2 is a prognostic biomarker and associated with immunotherapeutic responses in gastric cancer.” BMC cancer vol. 23,1 925. 2 Oct. 2023, doi:10.1186/s12885-023-11308-0
}
