% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{gene_table}
\alias{gene_table}
\title{Gene Expression Matrix}
\format{
A data frame with rows as genes and columns as samples.
}
\usage{
gene_table
}
\description{
A data.frame of expression values where rows correspond to genes and columns correspond to samples.
}
\keyword{datasets}
