% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{GGI.Sig}
\alias{GGI.Sig}
\title{Function to compute the signature as published by Sotiriou et al. in 2006.}
\usage{
GGI.Sig(expr, Response = NULL, verbose = FALSE, ...)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}

\item{...}{Other arguments passed on to ggi()'s params argument.}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Sotiriou et al. in 2006.
}
\examples{
a = GGI.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Sotiriou, Christos et al. “Gene expression profiling in breast cancer: understanding the molecular basis of histologic grade to improve prognosis.” Journal of the National Cancer Institute vol. 98,4 (2006): 262-72. doi:10.1093/jnci/djj052
}
