% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{GPS_3.Sig}
\alias{GPS_3.Sig}
\title{Function to compute the signature as published by Qi et al. in 2016.}
\usage{
GPS_3.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Qi et al. in 2016.
}
\examples{
a = GPS_3.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Qi, Lishuang et al. “An individualised signature for predicting response with concordant survival benefit for lung adenocarcinoma patients receiving platinum-based chemotherapy.” British journal of cancer vol. 115,12 (2016): 1513-1519. doi:10.1038/bjc.2016.370
}
