% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{HOT.score.Sig}
\alias{HOT.score.Sig}
\title{Function to compute the signature as published by Foy et al. in 2022.}
\usage{
HOT.score.Sig(expr, Response = NULL, verbose = FALSE, ...)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}

\item{...}{Other arguments passed on to gsva()'s params argument.}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Foy et al. in 2022.
}
\examples{
a = HOT.score.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Foy, Jean-Philippe et al. “Immunologically active phenotype by gene expression profiling is associated with clinical benefit from PD-1/PD-L1 inhibitors in real-world head and neck and lung cancer patients.” European journal of cancer (Oxford, England : 1990) vol. 174 (2022): 287-298. doi:10.1016/j.ejca.2022.06.034
}
