% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{IFN.Y.10.Sig}
\alias{IFN.Y.10.Sig}
\title{Function to compute the signature as published by Ayers et al. in 2017.}
\usage{
IFN.Y.10.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Ayers et al. in 2017.
}
\examples{
a = IFN.Y.10.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Ayers, Mark et al. “IFN-γ-related mRNA profile predicts clinical response to PD-1 blockade.” The Journal of clinical investigation vol. 127,8 (2017): 2930-2940. doi:10.1172/JCI91190
}
