% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{IFNy4gene.Sig}
\alias{IFNy4gene.Sig}
\title{Function to compute the signature as published by Higgs et al. in 2018.}
\usage{
IFNy4gene.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Higgs et al. in 2018.
}
\examples{
a = IFNy4gene.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Higgs, Brandon W et al. “Interferon Gamma Messenger RNA Signature in Tumor Biopsies Predicts Outcomes in Patients with Non-Small Cell Lung Carcinoma or Urothelial Cancer Treated with Durvalumab.” Clinical cancer research : an official journal of the American Association for Cancer Research vol. 24,16 (2018): 3857-3866. doi:10.1158/1078-0432.CCR-17-3451
}
