% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{ILB.Sig}
\alias{ILB.Sig}
\title{Function to compute the signature as published by Lv et al. in 2023.}
\usage{
ILB.Sig(expr, Response = NULL, verbose = FALSE, ...)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}

\item{...}{Other arguments passed on to gsva()'s params argument.}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Lv et al. in 2023.
}
\examples{
a = ILB.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Lv, Jiawei et al. “The tumor immune microenvironment of nasopharyngeal carcinoma after gemcitabine plus cisplatin treatment.” Nature medicine vol. 29,6 (2023): 1424-1436. doi:10.1038/s41591-023-02369-6
}
