% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{IPRES.Sig}
\alias{IPRES.Sig}
\title{Function to compute the signature as published by Hugo et al. in 2016.}
\usage{
IPRES.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Hugo et al. in 2016.
}
\examples{
a = IPRES.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Hugo, Willy et al. “Genomic and Transcriptomic Features of Response to Anti-PD-1 Therapy in Metastatic Melanoma.” Cell vol. 165,1 (2016): 35-44. doi:10.1016/j.cell.2016.02.065
}
