% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{Juul.2010.Sig}
\alias{Juul.2010.Sig}
\title{Function to compute the signature as published by Rooney et al. in 2015.}
\usage{
Juul.2010.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Rooney et al. in 2015.
}
\examples{
a = Juul.2010.Sig (expr = gene_table,Response = pdata_table$Response)

}
\references{
Rooney, Michael S et al. “Molecular and genetic properties of tumors associated with local immune cytolytic activity.” Cell vol. 160,1-2 (2015): 48-61. doi:10.1016/j.cell.2014.12.033
}
