% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{Liu.2018.Sig}
\alias{Liu.2018.Sig}
\title{Function to compute the signature as published by Liu et al. in 2018.}
\usage{
Liu.2018.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Liu et al. in 2018.
}
\examples{
a = Liu.2018.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Liu, Gang et al. “Seven Genes Based Novel Signature Predicts Clinical Outcome and Platinum Sensitivity of High Grade IIIc Serous Ovarian Carcinoma.” International journal of biological sciences vol. 14,14 2012-2022. 3 Nov. 2018, doi:10.7150/ijbs.28249
}
