% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{m6A.score.Sig}
\alias{m6A.score.Sig}
\title{Function to compute the signature as published by Zhang et al. in 2021.}
\usage{
m6A.score.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Zhang et al. in 2021.
}
\examples{
a = m6A.score.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Zhang, Zhihui et al. “m6A regulators as predictive biomarkers for chemotherapy benefit and potential therapeutic targets for overcoming chemotherapy resistance in small-cell lung cancer.” Journal of hematology & oncology vol. 14,1 190. 10 Nov. 2021, doi:10.1186/s13045-021-01173-4
}
