% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{MammaPrint.Sig}
\alias{MammaPrint.Sig}
\title{Function to compute the signature as published by van't Veer et al. in 2002.}
\usage{
MammaPrint.Sig(expr, Response = NULL, verbose = FALSE, ...)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}

\item{...}{Other arguments passed on to gene70()'s params argument.}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by van't Veer et al. in 2002.
}
\examples{
a = MammaPrint.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
van 't Veer, Laura J et al. “Gene expression profiling predicts clinical outcome of breast cancer.” Nature vol. 415,6871 (2002): 530-6. doi:10.1038/415530a
}
