% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{pdata_table}
\alias{pdata_table}
\title{Response Information}
\format{
A data frame with the sample names and their response data.
}
\usage{
pdata_table
}
\description{
A data frame containing the Response information for the samples.
}
\keyword{datasets}
