% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{PredictTax}
\alias{PredictTax}
\title{Predicting the sensitivity of cancer patients to Taxanes.}
\usage{
PredictTax(expr, Response = NULL, verbose = FALSE, ...)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step.}

\item{...}{Other arguments passed on to gsva()'s params argument.}
}
\description{
This function is based on gene set variation analysis (GSVA), which uses gene expression values to calculate the sensitivity score of samples to Taxanes.
}
\examples{
a = PredictTax(expr = gene_table,Response = pdata_table$Response)

}
