% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{RB.loss.Sig}
\alias{RB.loss.Sig}
\title{Function to compute the signature as published by Ertel et al. in 2010.}
\usage{
RB.loss.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Ertel et al. in 2010.
}
\examples{
a = RB.loss.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Ertel, Adam et al. “RB-pathway disruption in breast cancer: differential association with disease subtypes, disease-specific prognosis and therapeutic response.” Cell cycle (Georgetown, Tex.) vol. 9,20 (2010): 4153-63. doi:10.4161/cc.9.20.13454
}
