% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{RB.Sig}
\alias{RB.Sig}
\title{Function to compute the signature as published by Risi et al. in 2019.}
\usage{
RB.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Risi et al. in 2019.
}
\examples{
a = RB.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Risi, Emanuela et al. “An RB-1 loss of function gene signature as a tool to predict response to neoadjuvant chemotherapy plus anti-HER2 agents: a substudy of the NeoALTTO trial (BIG 1-06).” Therapeutic advances in medical oncology vol. 11 1758835919891608. 11 Dec. 2019, doi:10.1177/1758835919891608
}
