% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{Roh.immune.Sig}
\alias{Roh.immune.Sig}
\title{Function to compute the signature as published by Roh et al. in 2017.}
\usage{
Roh.immune.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Roh et al. in 2017.
}
\examples{
a = Roh.immune.Sig (expr = gene_table,Response = pdata_table$Response)

}
\references{
Roh, Whijae et al. “Integrated molecular analysis of tumor biopsies on sequential CTLA-4 and PD-1 blockade reveals markers of response and resistance.” Science translational medicine vol. 9,379 (2017): eaah3560. doi:10.1126/scitranslmed.aah3560
}
