% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{SELECT6.Sig}
\alias{SELECT6.Sig}
\title{Function to compute the signature as published by Lee et al. in 2021.}
\usage{
SELECT6.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Rooney et al. in 2021.
}
\examples{
a = SELECT6.Sig (expr = gene_table,Response = pdata_table$Response)

}
\references{
Lee, Joo Sang et al. “Synthetic lethality-mediated precision oncology via the tumor transcriptome.” Cell vol. 184,9 (2021): 2487-2502.e13. doi:10.1016/j.cell.2021.03.030
}
