% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{SIA.Sig}
\alias{SIA.Sig}
\title{Function to compute the signature as published by Mezheyeuski et al. in 2023.}
\usage{
SIA.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Mezheyeuski et al. in 2023.
}
\examples{
a = SIA.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Mezheyeuski, Artur et al. “An immune score reflecting pro- and anti-tumoural balance of tumour microenvironment has major prognostic impact and predicts immunotherapy response in solid cancers.” EBioMedicine vol. 88 (2023): 104452. doi:10.1016/j.ebiom.2023.104452
}
