% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{TIS.Sig}
\alias{TIS.Sig}
\title{Function to compute the signature as published by Şenbabaoğlu et al. in 2016.}
\usage{
TIS.Sig(expr, Response = NULL, verbose = FALSE, ...)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}

\item{...}{Other arguments passed on to gsva()'s params argument.}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Foy et al. in 2016.
}
\examples{
a = TIS.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Şenbabaoğlu, Yasin et al. “Tumor immune microenvironment characterization in clear cell renal cell carcinoma identifies prognostic and immunotherapeutically relevant messenger RNA signatures.” Genome biology vol. 17,1 231. 17 Nov. 2016, doi:10.1186/s13059-016-1092-z
}
