% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{TLS.Sig}
\alias{TLS.Sig}
\title{Function to compute the signature as published by Cabrita et al. in 2020.}
\usage{
TLS.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Cabrita et al. in 2020.
}
\examples{
a = TLS.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Cabrita, Rita et al. “Tertiary lymphoid structures improve immunotherapy and survival in melanoma.” Nature vol. 577,7791 (2020): 561-565. doi:10.1038/s41586-019-1914-8
}
