% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Signature.R
\name{Zhu.2020.Sig}
\alias{Zhu.2020.Sig}
\title{Function to compute the signature as published by Zhu et al. in 2020.}
\usage{
Zhu.2020.Sig(expr, Response = NULL, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{verbose}{Gives information about each calculation step}
}
\description{
This function computes signature scores from gene expression values following the algorithm used for the signature as published by Zhu et al. in 2020.
}
\examples{
a = Zhu.2020.Sig(expr = gene_table,Response = pdata_table$Response)

}
\references{
Zhu, Jieqiang et al. “Cancer genomics predicts disease relapse and therapeutic response to neoadjuvant chemotherapy of hormone sensitive breast cancers.” Scientific reports vol. 10,1 8188. 18 May. 2020, doi:10.1038/s41598-020-65055-4
}
