% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{calculateSig}
\alias{calculateSig}
\title{Function to compute the signatures.}
\usage{
calculateSig(expr, Response = NULL, Sig_list, verbose = FALSE)
}
\arguments{
\item{expr}{A matrix or data.frame of expression values where rows correspond to genes and columns correspond to samples.}

\item{Response}{A vector containing "R" and "NR" that matches the samples, used to calculate AUC. Default to NULL, and AUC will not be calculated.}

\item{Sig_list}{The signatures that need to be calculated.Setting the parameter to \code{all} means calculating all signatures.Setting the parameter to \code{Tax} or \code{ICI} means calculating all signatures for the corresponding category.}

\item{verbose}{Gives information about each calculation step.}
}
\description{
This function computes signatures scores from gene expression values.
}
\examples{
Sig_list = c('Docetaxel.Sig','DDIR.Sig','HOT.score.Sig','APM.Sig')
a = calculateSig(expr = gene_table,Response = pdata_table$Response,Sig_list = Sig_list)

}
